Introduce `.WallClock` module
This module collects API types and functions that are necessary to
support `BlockchainTime` API instances that depend on the wallclock (but
note that the `BlockchainTime` API _itself_ is entirely agnostic of the
real world).
