+++ ouroboros-consensus-shelley/src/Ouroboros/Consensus/Shelley/Ledger/Config.hs
@@ -15,7 +15,7 @@ import           Cardano.Prelude (NoUnexpectedThunks (..))
 import           Ouroboros.Network.Magic (NetworkMagic)
 
 import           Ouroboros.Consensus.Block
-import           Ouroboros.Consensus.BlockchainTime (SystemStart)
+import           Ouroboros.Consensus.BlockchainTime.WallClock
 import qualified Ouroboros.Consensus.HardFork.History as HardFork
 
 import qualified Shelley.Spec.Ledger.PParams as SL (ProtVer)
