+++ ouroboros-consensus/ouroboros-consensus.cabal
@@ -39,8 +39,8 @@ library
                        Ouroboros.Consensus.BlockchainTime
                        Ouroboros.Consensus.BlockchainTime.API
                        Ouroboros.Consensus.BlockchainTime.Simple
-                       Ouroboros.Consensus.BlockchainTime.SlotLength
                        Ouroboros.Consensus.BlockchainTime.SlotLengths
+                       Ouroboros.Consensus.BlockchainTime.WallClock
                        Ouroboros.Consensus.Config
                        Ouroboros.Consensus.Forecast
                        Ouroboros.Consensus.Fragment.Validated
