+++ ouroboros-consensus/src/Ouroboros/Consensus/BlockchainTime.hs
@@ -2,5 +2,5 @@ module Ouroboros.Consensus.BlockchainTime (module X) where
 
 import           Ouroboros.Consensus.BlockchainTime.API as X
 import           Ouroboros.Consensus.BlockchainTime.Simple as X
-import           Ouroboros.Consensus.BlockchainTime.SlotLength as X
 import           Ouroboros.Consensus.BlockchainTime.SlotLengths as X
+import           Ouroboros.Consensus.BlockchainTime.WallClock as X
