--- ouroboros-consensus/src/Ouroboros/Consensus/BlockchainTime/SlotLength.hs
@@ -1,52 +1,0 @@
-{-# LANGUAGE DeriveGeneric              #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
-
-module Ouroboros.Consensus.BlockchainTime.SlotLength (
-    -- * Slot length
-    SlotLength(..)
-    -- * Conversions
-  , slotLengthFromSec
-  , slotLengthToSec
-  , slotLengthFromMillisec
-  , slotLengthToMillisec
-  ) where
-
-import           Data.Fixed
-import           Data.Time
-import           GHC.Generics (Generic)
-
-import           Cardano.Prelude (NoUnexpectedThunks)
-
--- | Slot length
-newtype SlotLength = SlotLength { getSlotLength :: NominalDiffTime }
-  deriving (Show, Eq, Generic, NoUnexpectedThunks)
-
-{-------------------------------------------------------------------------------
-  Conversions
--------------------------------------------------------------------------------}
-
-slotLengthFromSec :: Integer -> SlotLength
-slotLengthFromSec = slotLengthFromMillisec . (* 1000)
-
-slotLengthToSec :: SlotLength -> Integer
-slotLengthToSec = (`div` 1000) . slotLengthToMillisec
-
-slotLengthFromMillisec :: Integer -> SlotLength
-slotLengthFromMillisec = SlotLength . conv
-  where
-    -- Explicit type annotation here means that /if/ we change the precision,
-    -- we are forced to reconsider this code.
-    conv :: Integer -> NominalDiffTime
-    conv = (realToFrac :: Pico -> NominalDiffTime)
-         . (/ 1000)
-         . (fromInteger :: Integer -> Pico)
-
-slotLengthToMillisec :: SlotLength -> Integer
-slotLengthToMillisec = conv . getSlotLength
-  where
-    -- Explicit type annotation here means that /if/ we change the precision,
-    -- we are forced to reconsider this code.
-    conv :: NominalDiffTime -> Integer
-    conv = truncate
-         . (* 1000)
-         . (realToFrac :: NominalDiffTime -> Pico)
