+++ ouroboros-consensus/src/Ouroboros/Consensus/Node/ErrorPolicy.hs
@@ -16,8 +16,7 @@ import           Ouroboros.Consensus.Storage.ImmutableDB.Types
                      (ImmutableDBError)
 import           Ouroboros.Consensus.Storage.VolatileDB.Types (VolatileDBError)
 
-import           Ouroboros.Consensus.BlockchainTime.Simple
-                     (SystemClockMovedBackException)
+import           Ouroboros.Consensus.BlockchainTime.WallClock
 import           Ouroboros.Consensus.MiniProtocol.BlockFetch.Server
                      (BlockFetchServerException)
 import           Ouroboros.Consensus.MiniProtocol.ChainSync.Client
