+++ ouroboros-consensus/src/Ouroboros/Consensus/Node/Run.hs
@@ -27,7 +27,7 @@ import           Ouroboros.Network.Magic (NetworkMagic)
 import           Ouroboros.Network.Protocol.LocalStateQuery.Codec (Some (..))
 
 import           Ouroboros.Consensus.Block
-import           Ouroboros.Consensus.BlockchainTime (SystemStart)
+import           Ouroboros.Consensus.BlockchainTime.WallClock
 import           Ouroboros.Consensus.Config
 import           Ouroboros.Consensus.HeaderValidation
 import           Ouroboros.Consensus.Ledger.Abstract
