+++ ouroboros-consensus/src/Ouroboros/Consensus/Node/Tracers.hs
@@ -23,8 +23,8 @@ import           Ouroboros.Network.TxSubmission.Outbound
                      (TraceTxSubmissionOutbound)
 
 import           Ouroboros.Consensus.Block (Header)
+import           Ouroboros.Consensus.BlockchainTime.WallClock
 import           Ouroboros.Consensus.Forecast (OutsideForecastRange)
-import           Ouroboros.Consensus.BlockchainTime (TraceBlockchainTimeEvent)
 import           Ouroboros.Consensus.Ledger.SupportsProtocol
 import           Ouroboros.Consensus.Mempool.API (ApplyTxErr, GenTx, GenTxId,
                      MempoolSize, TraceEventMempool)
