+++ ouroboros-consensus/test-consensus/Test/Consensus/HardFork/Infra.hs
@@ -30,7 +30,7 @@ import           Test.QuickCheck
 
 import           Cardano.Slotting.Slot
 
-import           Ouroboros.Consensus.BlockchainTime.SlotLength
+import           Ouroboros.Consensus.BlockchainTime.WallClock
 import qualified Ouroboros.Consensus.HardFork.History as HF
 import           Ouroboros.Consensus.Util.Counting
